requirements.txt
configparser
kafka-python